move file-collector to libKitsunemimiSakuraParser
